+++ CHANGELOG.md
@@ -81,3 +81,21 @@ Initial 1.16.3 release.
 ### Fixed
 - Mod now requires correct forge & mc versions
 - Should also work on 1.16.4
+
+## [2.0.2] - 2020-12-02
+### Fixed
+- Wrong server config descriptions
+.
+
+## [2.1.0] - 2020-12-02
+### Changed
+- Version range to allow usage with 1.16.1 & 1.16.2 (experimental).
+
+## [2.1.1] - 2020-12-02
+### Changed
+- FML Loader version to allow 32.*.
+
+## [2.1.2] - 2020-12-03
+### Fixed
+- Brazier runes not rendering for more than 9 blocks of the side
+- Rendering code being a mess